config/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-ckeditor==6.7.0
django-js-asset==2.1.0
idna==3.4
Pillow==10.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
